Atualização do client

+++ gframe/data_manager.h
@@ -148,7 +148,7 @@ public:
 		return _accessibilityTipsStrings.GetLocale(code);
 	}
 	std::vector<uint16_t> GetSetCode(const std::vector<std::wstring>& setname) const;
-	std::wstring GetNumString(int num, bool bracket = false) const;
+	std::wstring GetNumString(size_t num, bool bracket = false) const;
 	epro::wstringview FormatLocation(uint32_t location, int sequence) const;
 	std::wstring FormatAttribute(uint32_t attribute) const;
 	std::wstring FormatRace(uint64_t race, bool isSkill = false) const;
